+++ docs/changes.rst
@@ -24,6 +24,7 @@ v2.4.1 (UNRELEASED)
 - Update: ``cyanideandhappiness`` after site change.
 - Update: ``dilbert`` after site change.
 - Update: ``dustin`` after site change.
+- Update: ``hjalmar`` after file format change.
 - Update: ``nerfnow`` to include text.
 - Update: ``satw`` after site change.
 - Update: ``smbc`` after feed change.